feat(gallery):add-sort-to-people-tab
