+++ web/packages/new/photos/components/gallery/BarImpl.tsx
@@ -12,6 +12,10 @@ import { useIsSmallWidth } from "ente-base/components/utils/hooks";
 import { CollectionsSortOptions } from "ente-new/photos/components/CollectionsSortOptions";
 import { StarIcon } from "ente-new/photos/components/icons/StarIcon";
 import {
+    PeopleSortOptions,
+    type PeopleSortBy,
+} from "ente-new/photos/components/PeopleSortOptions";
+import {
     BarItemTile,
     ItemCard,
     TileTextOverlay,
@@ -39,8 +43,8 @@ import React, {
 import AutoSizer from "react-virtualized-auto-sizer";
 import {
     FixedSizeList,
-    type ListChildComponentProps,
     areEqual,
+    type ListChildComponentProps,
 } from "react-window";
 import { isMLSupported } from "../../services/ml";
 import type { GalleryBarMode } from "./reducer";
@@ -89,6 +93,14 @@ export interface GalleryBarImplProps {
      */
     onChangeCollectionsSortBy: (by: CollectionsSortBy) => void;
     /**
+     * The scheme that should be used for sorting the people.
+     */
+    peopleSortBy: PeopleSortBy;
+    /**
+     * Called when the user changes the people sorting scheme.
+     */
+    onChangePeopleSortBy: (by: PeopleSortBy) => void;
+    /**
      * The list of people that should be shown in the bar.
      */
     people: Person[];
@@ -112,6 +124,8 @@ export const GalleryBarImpl: React.FC<GalleryBarImplProps> = ({
     onShowAllPeople,
     collectionsSortBy,
     onChangeCollectionsSortBy,
+    peopleSortBy,
+    onChangePeopleSortBy,
     people,
     activePerson,
     onSelectPerson,
@@ -228,10 +242,16 @@ export const GalleryBarImpl: React.FC<GalleryBarImplProps> = ({
             direction="row"
             sx={{ alignItems: "center", gap: 1, minHeight: "64px" }}
         >
-            {mode != "people" && (
+            {mode != "people" ? (
                 <CollectionsSortOptions
                     activeSortBy={collectionsSortBy}
                     onChangeSortBy={onChangeCollectionsSortBy}
+                    transparentTriggerButtonBackground
+                />
+            ) : (
+                <PeopleSortOptions
+                    activeSortBy={peopleSortBy}
+                    onChangeSortBy={onChangePeopleSortBy}
                     transparentTriggerButtonBackground
                 />
             )}
@@ -248,10 +268,15 @@ export const GalleryBarImpl: React.FC<GalleryBarImplProps> = ({
             direction="row"
             sx={{ alignItems: "center", gap: 1, height: "64px" }}
         >
-            {mode != "people" && (
+            {mode != "people" ? (
                 <CollectionsSortOptions
                     activeSortBy={collectionsSortBy}
                     onChangeSortBy={onChangeCollectionsSortBy}
+                />
+            ) : (
+                <PeopleSortOptions
+                    activeSortBy={peopleSortBy}
+                    onChangeSortBy={onChangePeopleSortBy}
                 />
             )}
             <FilledIconButton
